Fixed AF and SB splitting; also some javadoc I should have done in the past and some refactoring I should have done in the past
