+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/mutect/filtering/MappingQualityFilter.java
@@ -22,14 +22,14 @@ public class MappingQualityFilter extends HardAlleleFilter {
     public ErrorType errorType() { return ErrorType.ARTIFACT; }
 
     @Override
-    public List<Boolean> areAllelesArtifacts(final VariantContext vc, final Mutect2FilteringEngine filteringEngine, ReferenceContext referenceContext) {
+    public List<Boolean> areAllelesArtifacts(final VariantContext vc, final Mutect2FilteringEngine filteringEngine, final ReferenceContext referenceContext) {
         final List<Integer> indelLengths = vc.getIndelLengths(); // alts only
-        List<Integer> mappingQualityByAllele = vc.getAttributeAsIntList(GATKVCFConstants.MEDIAN_MAPPING_QUALITY_KEY, 0);
+        final List<Integer> mappingQualityByAllele = vc.getAttributeAsIntList(GATKVCFConstants.MEDIAN_MAPPING_QUALITY_KEY, 0);
 
         // we use the mapping quality annotation of the alt allele in most cases, but for long indels we use the reference
         // annotation.  We have to do this because the indel, even if it maps uniquely, gets a poor mapping quality
         // by virtue of its mismatch.  The reference mapping quality is a decent proxy for the region's mappability.
-        int refQual = mappingQualityByAllele.remove(0); // get the ref value and convert list to alts only
+        final int refQual = mappingQualityByAllele.remove(0); // get the ref value and convert list to alts only
         new IndexRange(0, mappingQualityByAllele.size()).forEach(i -> {
             if (indelLengths != null && indelLengths.get(i) >= longIndelSize) {
                 mappingQualityByAllele.set(i, refQual);
