+++ src/test/java/org/broadinstitute/hellbender/tools/walkers/genotyper/AlleleSubsettingUtilsUnitTest.java
@@ -2,9 +2,12 @@ package org.broadinstitute.hellbender.tools.walkers.genotyper;
 
 import htsjdk.variant.variantcontext.*;
 import htsjdk.variant.vcf.VCFConstants;
+import htsjdk.variant.vcf.VCFHeader;
+import org.broadinstitute.hellbender.exceptions.UserException;
 import org.broadinstitute.hellbender.utils.MathUtils;
 import org.broadinstitute.hellbender.GATKBaseTest;
 import org.broadinstitute.hellbender.testutils.VariantContextTestUtils;
+import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.variant.GATKVCFConstants;
 import org.broadinstitute.hellbender.utils.variant.GATKVariantContextUtils;
 import org.testng.Assert;
@@ -15,6 +18,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 
 public class AlleleSubsettingUtilsUnitTest extends GATKBaseTest {
@@ -22,6 +26,112 @@ public class AlleleSubsettingUtilsUnitTest extends GATKBaseTest {
     private static final Allele Aref = Allele.create("A", true);
     private static final Allele C = Allele.create("C");
     private static final Allele G = Allele.create("G");
+
+    @DataProvider(name = "getIndexesOfRelevantAllelesData")
+    public Object[][] makeGetIndexesOfRelevantAllelesData() {
+        final int totalAlleles = 5;
+        final List<Allele> alleles = new ArrayList<>(totalAlleles);
+        alleles.add(Allele.create("A", true));
+        for ( int i = 1; i < totalAlleles; i++ )
+            alleles.add(Allele.create(Utils.dupChar('A', i + 1), false));
+
+        final List<Object[]> tests = new ArrayList<>();
+
+        for ( int alleleIndex = 0; alleleIndex < totalAlleles; alleleIndex++ ) {
+            tests.add(new Object[]{alleleIndex, alleles, true});
+            tests.add(new Object[]{alleleIndex, alleles, false});
+        }
+
+        return tests.toArray(new Object[][]{});
+    }
+
+    @DataProvider(name = "getIndexesOfRelevantAllelesDataSpanningDels")
+    public Object[][] makeGetIndexesOfRelevantAllelesDataSpanningDels() {
+        final int totalAlleles = 5;
+        final List<Allele> alleles = new ArrayList<>(totalAlleles);
+        alleles.add(Allele.create("A", true));
+        alleles.add(Allele.create("*", false));
+        alleles.add(Allele.create("*", false));
+        alleles.add(Allele.create("*", false));
+        alleles.add(Allele.NON_REF_ALLELE);
+
+        final List<Allele> suballeles = new ArrayList<>();
+        suballeles.add(Allele.create("A", true));
+        suballeles.add(Allele.create("*", false));
+
+        Genotype firstAltBest = new GenotypeBuilder("sampleName").alleles(suballeles).PL(new double[]{0, 0, 30, 0, 0, 20, 0, 0, 0, 10,
+                0, 0, 0, 0, 100, 0, 0, 0, 0, 0, 0}).make();
+        Genotype secondAltBest = new GenotypeBuilder("sampleName").alleles(suballeles).PL(new double[]{0, 0, 20, 0, 0, 30, 0, 0, 0, 10,
+                0, 0, 0, 0, 100, 0, 0, 0, 0, 0, 0}).make();
+        Genotype thirdAltBest = new GenotypeBuilder("sampleName").alleles(suballeles).PL(new double[]{0, 0, 20, 0, 0, 10, 0, 0, 0, 30,
+                0, 0, 0, 0, 100, 0, 0, 0, 0, 0, 0}).make();
+        Genotype altsTied = new GenotypeBuilder("sampleName").alleles(suballeles).PL(new double[]{0, 0, 20, 0, 0, 30, 0, 0, 0, 30,
+                0, 0, 0, 0, 100, 0, 0, 0, 0, 0, 0}).make();
+
+        final List<Object[]> tests = new ArrayList<>();
+
+        tests.add(new Object[]{alleles.stream().distinct().collect(Collectors.toList()), alleles, firstAltBest, 1});
+        tests.add(new Object[]{alleles.stream().distinct().collect(Collectors.toList()), alleles, secondAltBest, 2});
+        tests.add(new Object[]{alleles.stream().distinct().collect(Collectors.toList()), alleles, thirdAltBest, 3});
+        tests.add(new Object[]{alleles.stream().distinct().collect(Collectors.toList()), alleles, altsTied, 2});
+
+        return tests.toArray(new Object[][]{});
+    }
+
+    @Test(expectedExceptions = UserException.class)
+    public void testGetIndexesOfRelevantAllelesWithNoALT() {
+        final List<Allele> alleles1 = new ArrayList<>(1);
+        alleles1.add(Allele.create("A", true));
+        final List<Allele> alleles2 = new ArrayList<>(1);
+        alleles2.add(Allele.create("A", true));
+        GenotypeBuilder builder = new GenotypeBuilder();
+        AlleleSubsettingUtils.getIndexesOfRelevantAllelesForGVCF(alleles1, alleles2, -1, builder.make(), false);
+        AlleleSubsettingUtils.getIndexesOfRelevantAllelesForGVCF(alleles1, alleles2, -1, builder.make(), true);
+    }
+
+    @Test(dataProvider = "getIndexesOfRelevantAllelesData")
+    public void testGetIndexesOfRelevantAlleles(final int allelesIndex, final List<Allele> allAlleles, final boolean isSomatic) {
+        final List<Allele> myAlleles = new ArrayList<>(3);
+
+        // always add the reference and <NON_REF> alleles
+        myAlleles.add(allAlleles.get(0));
+        myAlleles.add(Allele.NON_REF_ALLELE);
+        // optionally add another alternate allele
+        if ( allelesIndex > 0 )
+            myAlleles.add(allAlleles.get(allelesIndex));
+
+        GenotypeBuilder builder = new GenotypeBuilder();
+
+        final int[] indexes = AlleleSubsettingUtils.getIndexesOfRelevantAllelesForGVCF(myAlleles, allAlleles, -1, builder.make(), isSomatic);
+
+        Assert.assertEquals(indexes.length, allAlleles.size());
+
+        for ( int i = 0; i < allAlleles.size(); i++ ) {
+            if ( i == 0 )
+                Assert.assertEquals(indexes[i], 0);    // ref should always match
+            else if ( i == allelesIndex )
+                Assert.assertEquals(indexes[i], 2);    // allele
+            else
+                Assert.assertEquals(indexes[i], 1);    // <NON_REF>
+        }
+    }
+
+    // This test asserts that when we us getINdexesOfRelevantAlleles in the case where there are multiple spanning deletions
+    // that we remap the PL indexes according to the BEST spanning deletion instead of the first one, which can happen if
+    // there were multiple spanning deletion alleles which are replaced with the same symbolic alleles before being fed to
+    // referenceConfidenceVariantContextMerger.
+    @Test (dataProvider = "getIndexesOfRelevantAllelesDataSpanningDels")
+    public void testGetIndexesOfRelevantAllelesMultiSpanningDel(final List<Allele> allelesToFind, final List<Allele> allAlleles, final Genotype g, final int expectedIndex) {
+        final boolean isSomatic = false; //Mutect2 doesn't output spanning deletions, so that's irrelevant
+        final int[] indexes = AlleleSubsettingUtils.getIndexesOfRelevantAllelesForGVCF(allAlleles, allelesToFind,-1, g, isSomatic);
+
+        Assert.assertEquals(indexes.length, allelesToFind.size());
+
+        // Asserting that the expected index for the spanning deletion allele corresponds to the most likely one according to the PL
+        Assert.assertEquals(indexes[0], 0);    // ref should always match
+        Assert.assertEquals(indexes[1], expectedIndex);    // allele
+        Assert.assertEquals(indexes[2], 4);    // <ALT>
+    }
 
     @Test(dataProvider = "updatePLsSACsAndADData")
     public void testUpdatePLsAndADData(final VariantContext originalVC,
