+++ src/test/java/org/broadinstitute/hellbender/tools/walkers/mutect/Mutect2IntegrationTest.java
@@ -571,7 +571,7 @@ public class Mutect2IntegrationTest extends CommandLineProgramTest {
                 .map(VariantContext::getFilters).collect(Collectors.toList());
 
         final List<List<String>> actualASFilters = VariantContextTestUtils.streamVcf(filteredVcf)
-                .map(vc -> AnnotationUtils.decodeAnyASListWithPrintDelim(vc.getCommonInfo().getAttributeAsString(GATKVCFConstants.AS_FILTER_STATUS_KEY, ""))).collect(Collectors.toList());
+                .map(vc -> AnnotationUtils.decodeAnyASListWithRawDelim(vc.getCommonInfo().getAttributeAsString(GATKVCFConstants.AS_FILTER_STATUS_KEY, ""))).collect(Collectors.toList());
         Assert.assertEquals(actualASFilters, expectedASFilters);
 
         Assert.assertEquals(actualFilters.size(), expectedFilters.size());
